+++ src/stream_output.h
@@ -1,6 +1,6 @@
 /*
 	dtee - run a program with standard output and standard error copied to files
-	Copyright 2018,2021  Simon Arlott
+	Copyright 2018,2021,2024  Simon Arlott
 
 	This program is free software: you can redistribute it and/or modify
 	it under the terms of the GNU General Public License as published by
@@ -34,6 +34,7 @@ public:
 	bool output(OutputType type, const std::vector<char> &buffer, size_t len) override;
 
 private:
+	bool set_blocking(int fd, const char *name_fmt);
 	bool output(int fd, const char *name_fmt, const std::vector<char> &buffer, size_t len);
 };