+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Changed
   handlers when using Boost 1.82+.
 * Disable "error location" in exception messages when building with Boost.Asio
   1.79 because these exceptions are already caught and properly identified.
+* Configure standard output and standard error to be blocking.
 
 Fixed
 ~~~~~